Create Osuji_ Ifunanyachukwu Maryann 2024714101
A python program implements a grading system using 'if' and 'elif' statement. It takes a student score and print the corresponding based on the following scale:
. 70_100 Grade A
.60_69 Grade B
.50_59 Grade C
.49_45 Grade D
.44_40 Grade E
.0_39 Grade F
For each value score the program also prints "YES" to confirm that the grading was successful. If the score is outside the valid range . It prints "Invalid score" or "NO".
Created by : Osuji Ifunanyachukwu Maryann 
Reg no: 2024714101